backend-Flask/requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Script==2.0.6
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
password-strength==0.0.3.post2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
